fix: 승패 판정 및 추가 카드 수령 로직 오류 수정
