+++ src/main/java/blackjack/view/ResultView.java
@@ -7,7 +7,7 @@ public class ResultView {
 
 	private static final String MESSAGE_HAND_OUT_CARD = "%n딜러와 %s 에게 2장의 카드를 나누었습니다.%n";
 	private static final String NAME_DELIMITER = ", ";
-	private static final String MESSAGE_SCORE_OVER_21 = "가진 카드의 합이 21 이상이 되어 더 이상 카드를 받을 수 없습니다.";
+	private static final String MESSAGE_SCORE_OVER_21 = "가진 카드의 합이 21을 초과하여 더 이상 카드를 받을 수 없습니다.";
 	private static final String MESSAGE_DEALER_NOT_RECEIVE = "딜러는 16이하라 한장의 카드를 더 받았습니다";
 	private static final String MESSAGE_DEALER_RECEIVE = "딜러는 17이상이라 카드를 더 받지 않았습니다.";
 	private static final String RESULT_DELIMITER = " - 결과: ";
